page visibility

+++ src/main/java/com/issuetracker/pages/project/ProjectDetail.java
@@ -9,9 +9,9 @@ import com.issuetracker.model.TypeId;
 import com.issuetracker.pages.component.component.ComponentListView;
 import com.issuetracker.pages.component.customField.CustomFieldListView;
 import com.issuetracker.pages.component.version.VersionListView;
+import com.issuetracker.pages.permissions.AccessDenied;
 import com.issuetracker.pages.permissions.ProjectPermission;
 import com.issuetracker.pages.validator.ProjectNameValidator;
-import com.issuetracker.service.api.PermissionService;
 import com.issuetracker.service.api.ProjectService;
 import com.issuetracker.service.api.SecurityService;
 import static com.issuetracker.web.Constants.*;
@@ -40,18 +40,19 @@ import org.jboss.logging.Logger;
  */
 public class ProjectDetail extends PageLayout {
 
+    private final Logger log = Logger.getLogger(ProjectDetail.class);
+    
     @Inject
     private ProjectService projectService;
     @Inject
     private SecurityService securityService;
-    @Inject
-    private PermissionService permissionService;
     
     private final Form<String> renameForm;
     private final Form<ProjectVersion> projectVersionForm;
     private final Form<Component> projectComponentForm;
     private final Form<CustomField> projectCustomFieldForm;
     private Link deleteProject;
+    private Link projectPermissions;
     private List<ProjectVersion> projectVersionList;    
     private List<Component> projectComponentList;    
     private List<CustomField> projectCustomFieldsList;    
@@ -72,17 +73,27 @@ public class ProjectDetail extends PageLayout {
         projectComponentForm.setVisible(permittedActions.contains(roles.getProperty("it.project.components")));
         projectCustomFieldForm.setVisible(permittedActions.contains(roles.getProperty("it.project.custom.fields")));
         deleteProject.setVisible(permittedActions.contains(roles.getProperty("it.project.delete")));
+        projectPermissions.setVisible(permittedActions.contains(roles.getProperty("it.project.permissions")));
     }
     
     public ProjectDetail(PageParameters parameters) {
         StringValue stringProjectId = parameters.get("project");
         if (stringProjectId.equals(StringValue.valueOf((String)null))) {
+            log.warn("Page parameters doesn't contain project id. Redirecting to Home page.");
             throw new RedirectToUrlException(HOME_PAGE);
         }
         Long projectId = stringProjectId.toLong();
-        permittedActions = securityService.getPermittedActionsForUserAndItem(TypeId.project, projectId);
-        
         project = projectService.getProjectById(projectId);
+        
+        if (project == null) {
+            log.warn("Project with given id doesn't exist. Redirecting to Home page.");
+            throw new RedirectToUrlException(HOME_PAGE);
+        }
+        permittedActions = securityService.getPermittedActionsForUserAndItem(TypeId.project, projectId);
+
+        if (!permittedActions.contains(roles.getProperty("it.project.browse"))) {
+            setResponsePage(AccessDenied.class);
+        }       
         
         projectVersionList = project.getVersions();
         projectComponentList = project.getComponents();
@@ -93,6 +104,7 @@ public class ProjectDetail extends PageLayout {
         renameForm = new Form<String>("renameForm") {
             @Override
             protected void onSubmit() {
+                checkPermission("it.project.rename");
                 projectService.update(project);
             }
         };
@@ -104,20 +116,23 @@ public class ProjectDetail extends PageLayout {
 
             @Override
             public void onClick() {
+                checkPermission("it.project.delete");
                 projectService.remove(project);
                 setResponsePage(ListProjects.class);
             }
         };
         add(deleteProject);
         
-        add(new Link("permissions") {
+        projectPermissions = new Link("permissions") {
             @Override
             public void onClick() {
+                checkPermission("it.project.permissions");
                 PageParameters pageParameters = new PageParameters();
                 pageParameters.add("project", project.getId());
                 setResponsePage(ProjectPermission.class, pageParameters);
             }
-        });
+        };
+        add(projectPermissions);
         
 //<editor-fold defaultstate="collapsed" desc="versions">
         IModel<List<ProjectVersion>> versionModel = new CompoundPropertyModel<List<ProjectVersion>>(projectVersionList) {
@@ -137,6 +152,7 @@ public class ProjectDetail extends PageLayout {
         projectVersionForm = new Form<ProjectVersion>("projectVersionForm") {
             @Override
             protected void onSubmit() {
+                checkPermission("it.project.versions");
                 ProjectVersion projectVersion = new ProjectVersion();
                 projectVersion.setName(versionTextField.getInput());
                 stringVersion = null; //this will clear version textfield
@@ -167,6 +183,7 @@ public class ProjectDetail extends PageLayout {
         projectComponentForm = new Form<Component>("projectComponentForm") {
             @Override
             protected void onSubmit() {
+                checkPermission("it.project.components");
                 Component component = new Component();
                 component.setName(componentTextField.getInput());
                 stringComponent = null; //this will clear component textfield
@@ -197,6 +214,7 @@ public class ProjectDetail extends PageLayout {
         projectCustomFieldForm = new Form<CustomField>("projectCustomFieldForm") {
             @Override
             protected void onSubmit() {
+                checkPermission("it.project.custom.fields");
                 CustomField customField = new CustomField();
                 customField.setCfName(customFieldTextField.getInput());
                 stringCustomField = null; //this will clear custom field text field
@@ -208,23 +226,6 @@ public class ProjectDetail extends PageLayout {
         projectCustomFieldForm.add(customFieldTextField);
         add(projectCustomFieldForm);
 //</editor-fold>
-        
-//<editor-fold defaultstate="collapsed" desc="permissions">
-//        final List<PermissionToDisplay> defaultPermissions = permissionService.getDefaultPermissions();
-//        IModel<List<PermissionToDisplay>> permissionsModel = new CompoundPropertyModel<List<PermissionToDisplay>>(defaultPermissions) {
-//            @Override
-//            public List<PermissionToDisplay> getObject() {
-//                return defaultPermissions;            
-//            }
-//        };
-//        PermissionsListView<PermissionToDisplay> permissionsListView = new PermissionsListView<>("permissionsListView", permissionsModel, new Model<>(project));
-//        wmcPermission = new WebMarkupContainer("wmcPermission");
-//        wmcPermission.add(permissionsListView);
-//        wmcPermission.setOutputMarkupId(true);
-//        add(wmcPermission);
-//</editor-fold>
-        
-        Logger.getLogger(ProjectDetail.class).warn("TODO: link to Project's issues.");
     }
 
     public String getStringVersion() {
@@ -257,5 +258,11 @@ public class ProjectDetail extends PageLayout {
 
     public void setProject(Project project) {
         this.project = project;
+    }
+    
+    private void checkPermission(String name) {
+        if (!permittedActions.contains(roles.getProperty(name))) {
+            setResponsePage(AccessDenied.class);
+        }
     }
 }
